Refactor simple CLI commands (#7619)
