+++ packages/astro/src/cli/add/babel.ts
