+++ packages/astro/src/cli/docs/open.ts
