+++ packages/astro/src/cli/telemetry/index.ts
@@ -1,9 +1,9 @@
 /* eslint-disable no-console */
 import type yargs from 'yargs-parser';
-import * as msg from '../core/messages.js';
-import { telemetry } from '../events/index.js';
+import * as msg from '../../core/messages.js';
+import { telemetry } from '../../events/index.js';
 
-export interface TelemetryOptions {
+interface TelemetryOptions {
 	flags: yargs.Arguments;
 }
