+++ packages/astro/src/core/config/config.ts
@@ -7,7 +7,6 @@ import path from 'path';
 import { fileURLToPath, pathToFileURL } from 'url';
 import { mergeConfig as mergeViteConfig } from 'vite';
 import { AstroError, AstroErrorData } from '../errors/index.js';
-import type { LogOptions } from '../logger/core.js';
 import { arraify, isObject, isURL } from '../util.js';
 import { createRelativeSchema } from './schema.js';
 import { loadConfigWithVite } from './vite-load.js';
@@ -163,7 +162,6 @@ interface LoadConfigOptions {
 	flags?: Flags;
 	cmd: string;
 	validate?: boolean;
-	logging: LogOptions;
 	/** Invalidate when reloading a previously loaded config */
 	isRestart?: boolean;
 	fsMod?: typeof fs;
